Create Rental prices across US.R
We can see one very clear influential point where the average rent for a 2 bedroom apartment increased by more than \$600 over 2019. Upon further investigation, the region is Siesta Key in Florida. More specifically monthly rent increased by \$650. It makes sense that the most influential point to the maximum difference in rent over 2019 is the region with the largest difference. By removing this point, the maximum increase in rental price is $615. 

I can think of a couple of reasons why the average monthly rent increase in Siesta Key is the most influential point to the attribute max:

- Siesta Key is right by the beach along the coast of Florida, making it an area with incredible views and attractive real estate
- There is also a reality tv show called Siesta Key that takes place in the region which may have increased its publicity